// apps/snsgov/src/candid/sns_governance.did.d.ts
export interface Account {
  'owner' : [] | [Principal],
  'subaccount' : [] | [Subaccount],
}
export type Action = {
    'ManageNervousSystemParameters' : NervousSystemParameters
  } |
  { 'AddGenericNervousSystemFunction' : NervousSystemFunction } |
  { 'ManageDappCanisterSettings' : ManageDappCanisterSettings } |
  { 'ExecuteExtensionOperation' : ExecuteExtensionOperation } |
  { 'UpgradeExtension' : UpgradeExtension } |
  { 'RemoveGenericNervousSystemFunction' : bigint } |
  { 'SetTopicsForCustomProposals' : SetTopicsForCustomProposals } |
  { 'RegisterExtension' : RegisterExtension } |
  { 'UpgradeSnsToNextVersion' : {} } |
  { 'RegisterDappCanisters' : RegisterDappCanisters } |
  { 'TransferSnsTreasuryFunds' : TransferSnsTreasuryFunds } |
  { 'UpgradeSnsControlledCanister' : UpgradeSnsControlledCanister } |
  { 'DeregisterDappCanisters' : DeregisterDappCanisters } |
  { 'MintSnsTokens' : MintSnsTokens } |
  { 'AdvanceSnsTargetVersion' : AdvanceSnsTargetVersion } |
  { 'Unspecified' : {} } |
  { 'ManageSnsMetadata' : ManageSnsMetadata } |
  {
    'ExecuteGenericNervousSystemFunction' : ExecuteGenericNervousSystemFunction
  } |
  { 'ManageLedgerParameters' : ManageLedgerParameters } |
  { 'Motion' : Motion };
export type ActionAuxiliary = {
    'TransferSnsTreasuryFunds' : MintSnsTokensActionAuxiliary
  } |
  { 'MintSnsTokens' : MintSnsTokensActionAuxiliary } |
  { 'AdvanceSnsTargetVersion' : AdvanceSnsTargetVersionActionAuxiliary };
export interface AddNeuronPermissions {
  'permissions_to_add' : [] | [NeuronPermissionList],
  'principal_id' : [] | [Principal],
}
export interface AdvanceSnsTargetVersion { 'new_target' : [] | [SnsVersion] }
export interface AdvanceSnsTargetVersionActionAuxiliary {
  'target_version' : [] | [SnsVersion],
}
export interface Amount { 'e8s' : bigint }
export interface Ballot {
  'vote' : number,
  'cast_timestamp_seconds' : bigint,
  'voting_power' : bigint,
}
export type By = { 'MemoAndController' : MemoAndController } |
  { 'NeuronId' : {} };
// ...
export interface ClaimOrRefresh { 'by' : [] | [By] }
// ...
export type Command = { 'Split' : Split } |
  { 'Follow' : Follow } |
  { 'DisburseMaturity' : DisburseMaturity } |
  { 'ClaimOrRefresh' : ClaimOrRefresh } |
  { 'Configure' : Configure } |
  { 'RegisterVote' : RegisterVote } |
  { 'SetFollowing' : SetFollowing } |
  { 'MakeProposal' : Proposal } |
  { 'StakeMaturity' : StakeMaturity } |
  { 'RemoveNeuronPermissions' : RemoveNeuronPermissions } |
  { 'AddNeuronPermissions' : AddNeuronPermissions } |
  { 'MergeMaturity' : MergeMaturity } |
  { 'Disburse' : Disburse };
// ...
export interface Configure { 'operation' : [] | [Operation] }
// ...
export interface Disburse {
  'to_account' : [] | [Account],
  'amount' : [] | [Amount],
}
export interface DisburseMaturity {
  'to_account' : [] | [Account],
  'percentage_to_disburse' : number,
}
// ...
export interface Follow {
  'function_id' : bigint,
  'followees' : Array<NeuronId>,
}
export interface Followee {
  'alias' : [] | [string],
  'neuron_id' : [] | [NeuronId],
}
export interface Followees { 'followees' : Array<NeuronId> }
export interface FolloweesForTopic {
  'topic' : [] | [Topic],
  'followees' : Array<Followee>,
}
// ...
export interface MergeMaturity { 'percentage_to_merge' : number }
// ...
export interface NeuronId { 'id' : Uint8Array }
export interface NeuronIds { 'neuron_ids' : Array<NeuronId> }
export interface NeuronInFlightCommand {
  'command' : [] | [Command_2],
  'timestamp' : bigint,
}
export interface NeuronPermission {
  'principal' : [] | [Principal],
  'permission_type' : Int32Array,
}
export interface NeuronPermissionList { 'permissions' : Int32Array }
// ...
export type Operation = {
    'ChangeAutoStakeMaturity' : ChangeAutoStakeMaturity
  } |
  { 'StopDissolving' : {} } |
  { 'StartDissolving' : {} } |
  { 'IncreaseDissolveDelay' : IncreaseDissolveDelay } |
  { 'SetDissolveTimestamp' : SetDissolveTimestamp };
// ...
export interface Proposal {
  'url' : string,
  'title' : string,
  'action' : [] | [Action],
  'summary' : string,
}
// ...
export interface ProposalId { 'id' : bigint }
// ...
export interface RegisterVote {
  'vote' : number,
  'proposal' : [] | [ProposalId],
}
export interface RegisteredExtensionOperationSpec {
  'spec' : [] | [ExtensionOperationSpec],
  'canister_id' : [] | [Principal],
}
export interface RemoveNeuronPermissions {
  'permissions_to_remove' : [] | [NeuronPermissionList],
  'principal_id' : [] | [Principal],
}
// ...
export interface SetFollowing { 'topic_following' : Array<FolloweesForTopic> }
// ...
export interface Split { 'memo' : bigint, 'amount_e8s' : bigint }
export interface SplitResponse { 'created_neuron_id' : [] | [NeuronId] }
export interface StakeMaturity { 'percentage_to_stake' : [] | [number] }